requirements.txt
anyio==4.11.0 ; python_version >= "3.13" and python_version < "4"
blinker==1.9.0 ; python_version >= "3.13" and python_version < "4"
cachetools==6.2.1 ; python_version >= "3.13" and python_version < "4"
certifi==2025.10.5 ; python_version >= "3.13" and python_version < "4"
charset-normalizer==3.4.4 ; python_version >= "3.13" and python_version < "4"
click==8.3.0 ; python_version >= "3.13" and python_version < "4"
cloudevents==1.11.0 ; python_version >= "3.13" and python_version < "4"
colorama==0.4.6 ; python_version >= "3.13" and python_version < "4" and platform_system == "Windows"
db-dtypes==1.4.3 ; python_version >= "3.13" and python_version < "4"
deprecation==2.1.0 ; python_version >= "3.13" and python_version < "4"
flask==3.1.2 ; python_version >= "3.13" and python_version < "4"
functions-framework==3.9.2 ; python_version >= "3.13" and python_version < "4"
google-api-core==2.28.1 ; python_version >= "3.13" and python_version < "4"
google-auth-oauthlib==1.2.2 ; python_version >= "3.13" and python_version < "4"
google-auth==2.43.0 ; python_version >= "3.13" and python_version < "4"
google-cloud-bigquery==3.38.0 ; python_version >= "3.13" and python_version < "4"
google-cloud-core==2.5.0 ; python_version >= "3.13" and python_version < "4"
google-crc32c==1.7.1 ; python_version >= "3.13" and python_version < "4"
google-resumable-media==2.7.2 ; python_version >= "3.13" and python_version < "4"
googleapis-common-protos==1.72.0 ; python_version >= "3.13" and python_version < "4"
grpcio-status==1.76.0 ; python_version >= "3.13" and python_version < "4"
grpcio==1.76.0 ; python_version >= "3.13" and python_version < "4"
gunicorn==23.0.0 ; python_version >= "3.13" and python_version < "4"
h11==0.16.0 ; python_version >= "3.13" and python_version < "4"
idna==3.11 ; python_version >= "3.13" and python_version < "4"
itsdangerous==2.2.0 ; python_version >= "3.13" and python_version < "4"
jinja2==3.1.6 ; python_version >= "3.13" and python_version < "4"
markupsafe==3.0.3 ; python_version >= "3.13" and python_version < "4"
numpy==2.3.4 ; python_version >= "3.13" and python_version < "4"
oauthlib==3.3.1 ; python_version >= "3.13" and python_version < "4"
packaging==25.0 ; python_version >= "3.13" and python_version < "4"
pandas-gbq==0.30.0 ; python_version >= "3.13" and python_version < "4"
pandas==2.3.3 ; python_version >= "3.13" and python_version < "4"
proto-plus==1.26.1 ; python_version >= "3.13" and python_version < "4"
protobuf==6.33.0 ; python_version >= "3.13" and python_version < "4"
pyarrow==22.0.0 ; python_version >= "3.13" and python_version < "4"
pyasn1-modules==0.4.2 ; python_version >= "3.13" and python_version < "4"
pyasn1==0.6.1 ; python_version >= "3.13" and python_version < "4"
pydata-google-auth==1.9.1 ; python_version >= "3.13" and python_version < "4"
python-dateutil==2.9.0.post0 ; python_version >= "3.13" and python_version < "4"
python-dotenv==1.2.1 ; python_version >= "3.13" and python_version < "4"
pytz==2025.2 ; python_version >= "3.13" and python_version < "4"
requests-oauthlib==2.0.0 ; python_version >= "3.13" and python_version < "4"
requests==2.32.5 ; python_version >= "3.13" and python_version < "4"
rsa==4.9.1 ; python_version >= "3.13" and python_version < "4"
setuptools==80.9.0 ; python_version >= "3.13" and python_version < "4"
six==1.17.0 ; python_version >= "3.13" and python_version < "4"
sniffio==1.3.1 ; python_version >= "3.13" and python_version < "4"
starlette==0.50.0 ; python_version >= "3.13" and python_version < "4"
typing-extensions==4.15.0 ; python_version >= "3.13" and python_version < "4"
tzdata==2025.2 ; python_version >= "3.13" and python_version < "4"
urllib3==2.5.0 ; python_version >= "3.13" and python_version < "4"
uvicorn-worker==0.4.0 ; python_version >= "3.13" and python_version < "4"
uvicorn==0.38.0 ; python_version >= "3.13" and python_version < "4"
watchdog==6.0.0 ; python_version >= "3.13" and python_version < "4"
werkzeug==3.1.3 ; python_version >= "3.13" and python_version < "4"
